djangoProject/requirements.txt
djangorestframework
markdown
django-filter
django-cors-headers
django-mysql
PyMySql
numpy